dax_tag_handle() fixes

+++ lib/libdax.h
@@ -87,7 +87,7 @@ int opt_get_msgtimeout(void);
 typedef struct CDT_Member {
     char *name;
     unsigned int type;
-    size_t count;
+    u_int32_t count;
     struct CDT_Member *next;
 } cdt_member;
